# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Port Logistics API",
        "description": "Port Logistics API",
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://port_logistic.com/",
            "description": ""
        }
    ],
    "paths": {
...
        "/send-email": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "User",
                "operationId": "App\\Http\\Controllers\\API\\V1\\SendEmailController::sendEmail",
                "requestBody": {
                    "description": "Pass email details for send email",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "name",
                                    "email",
                                    "template"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "format": "number",
                                        "example": 1
                                    },
                                    "name": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "test"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "test@gmail.com"
                                    },
                                    "template": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "register"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and mobile to get the authentication token",
                "name": "Token based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}